Switch to format 3.0 (quilt)
It has happened several times to me that I merged a new upstream
release without bumping debian/changelog or vice versa and only
noticed this problem when I ran lintian.  In the 3.0 (quilt) format,
dpkg-source catches this mistake early.

The only advantage of the 1.0 format is that it is easier to
cherry-pick changes from upstream if they also use git, which is not
the case with xterm.