do_select_regex: Support matching before the cursor
With the old code the cursor had to be before the last character of the
match. This way you had to move back at least one character, before
triggering functions like exec-selectable. Including the character at the
cursor in the match makes it easier to type something and then hit the key
to trigger exec-selectable, e.g. to open the browser.