exec-/insert-selectable: use pointer on mouse clicks
If exec-selectable or insert-selection gets triggered by a mouse button
click, the position of the mouse pointer should be used. This makes it
easier to address any position on the window and it makes it possible to
use the mouse, if you can't move the cursor in applications like mutt.

An example:

mutt.*.VT100.translations: #override\
  Alt <Btn1Up>: exec-selectable("firefox '%s'", "regex ^https?://[^ ]+[^>[:space:]]")\n\
  @ISO_Level3_Shift Shift <Key> w: exec-selectable("firefox '%s'", "regex ^https?://[^ ]+[^>[:space:]]")